requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
Werkzeug==2.3.3
python-dotenv==0.19.2
gunicorn==20.1.0
paho-mqtt==1.6.1
bcrypt==4.2.0
